+++ retinazer-gwt-test/src/test/java/com/github/antag99/retinazer/GWTEntityProcessorSystemTest.java
@@ -15,7 +15,7 @@ public class GWTEntityProcessorSystemTest extends RetinazerTestCase {
         protected void process(int entity) {
             if (processedEntities.contains(entity))
                 throw new IllegalStateException("entity already processed: " + entity);
-            processedEntities.addEntity(entity);
+            processedEntities.edit().addEntity(entity);
         }
     }
 
@@ -29,16 +29,16 @@ public class GWTEntityProcessorSystemTest extends RetinazerTestCase {
         entities.set(c = engine.createEntity().idx());
         engine.update();
         assertEquals(system.processedEntities.getMask(), entities);
-        system.processedEntities.clear();
+        system.processedEntities.edit().clear();
         engine.destroyEntity(b);
         entities.clear(b);
         engine.update();
         assertEquals(system.processedEntities.getMask(), entities);
-        system.processedEntities.clear();
+        system.processedEntities.edit().clear();
         entities.set(b = engine.createEntity().idx());
         engine.update();
         assertEquals(system.processedEntities.getMask(), entities);
-        system.processedEntities.clear();
+        system.processedEntities.edit().clear();
         engine.destroyEntity(a);
         engine.destroyEntity(b);
         engine.destroyEntity(c);