+++ retinazer-gwt-test/src/test/java/com/github/antag99/retinazer/beam/system/GWTSpatialSystemTest.java
@@ -54,7 +54,7 @@ public final class GWTSpatialSystemTest extends RetinazerTestCase {
         Handle entity = engine.createEntity().cpy();
 
         EntitySet set = new EntitySet();
-        set.addEntity(entity.idx());
+        set.edit().addEntity(entity.idx());
 
         Position position = entity.create(Position.class);
         position.x = 0f;