+++ retinazer/src/test/java/com/github/antag99/retinazer/EntitySetTest.java
@@ -6,28 +6,25 @@ import java.util.Arrays;
 
 import org.junit.Test;
 
-import com.badlogic.gdx.utils.IntArray;
 import com.github.antag99.retinazer.util.Mask;
 
 public class EntitySetTest {
-    public static final class TestEntitySetListener implements EntitySetListener {
+    public static final class TestEntitySetListener implements EntityListener {
         public Mask inserted = new Mask();
         public Mask removed = new Mask();
         public int insertedInvocations = 0;
         public int removedInvocations = 0;
 
         @Override
-        public void inserted(IntArray entities) {
+        public void inserted(EntitySet entities) {
             insertedInvocations++;
-            for (int i = 0, n = entities.size; i < n; i++)
-                inserted.set(entities.get(i));
+            inserted.or(entities.getMask());
         }
 
         @Override
-        public void removed(IntArray entities) {
+        public void removed(EntitySet entities) {
             removedInvocations++;
-            for (int i = 0, n = entities.size; i < n; i++)
-                removed.set(entities.get(i));
+            removed.or(entities.getMask());
         }
 
         public void verifyInserted(int... entities) {
@@ -74,104 +71,14 @@ public class EntitySetTest {
     }
 
     @Test
-    public void testAddListener() {
-        TestEntitySetListener testListener = new TestEntitySetListener();
-        EntitySet entitySet = new EntitySet();
-        entitySet.addListener(testListener);
-
-        int entity = 0;
-
-        entitySet.addEntity(entity);
-
-        testListener.verifyInserted(entity);
-        testListener.verifyRemoved(new int[0]);
-        testListener.verifyInsertedInvocations(1);
-        testListener.verifyRemovedInvocations(0);
-
-        entitySet.removeEntity(0);
-
-        testListener.verifyInserted(new int[0]);
-        testListener.verifyRemoved(entity);
-        testListener.verifyInsertedInvocations(0);
-        testListener.verifyRemovedInvocations(1);
-
-        // Re-inserts the listener at index 0
-        entitySet.addListener(testListener);
-        entitySet.addEntity(0);
-
-        testListener.verifyInserted(entity);
-        testListener.verifyRemoved(new int[0]);
-        testListener.verifyInsertedInvocations(1);
-        testListener.verifyRemovedInvocations(0);
-
-        entitySet.removeEntity(0);
-
-        testListener.verifyInserted(new int[0]);
-        testListener.verifyRemoved(entity);
-        testListener.verifyInsertedInvocations(0);
-        testListener.verifyRemovedInvocations(1);
-    }
-
-    @Test
-    public void testRemoveListener() {
-        TestEntitySetListener testListener = new TestEntitySetListener();
-        TestEntitySetListener testListener2 = new TestEntitySetListener();
-        EntitySet entitySet = new EntitySet();
-        // Removing the listener when it does not exist has no effect
-        entitySet.removeListener(testListener);
-        entitySet.addListener(testListener);
-        entitySet.addListener(testListener);
-        entitySet.addListener(testListener);
-        entitySet.addListener(testListener2);
-        entitySet.removeListener(testListener);
-        // One listener; duplicate listeners are ignored
-
-        // Paranoia!
-        testListener.verifyInserted(new int[0]);
-        testListener.verifyRemoved(new int[0]);
-        testListener.verifyInsertedInvocations(0);
-        testListener.verifyRemovedInvocations(0);
-
-        int entity = 0;
-
-        entitySet.addEntity(entity);
-
-        // Shouldn't be invoked
-        testListener.verifyInserted(new int[0]);
-        testListener.verifyRemoved(new int[0]);
-        testListener.verifyInsertedInvocations(0);
-        testListener.verifyRemovedInvocations(0);
-
-        testListener2.verifyInserted(entity);
-        testListener2.verifyRemoved(new int[0]);
-        testListener2.verifyInsertedInvocations(1);
-        testListener2.verifyRemovedInvocations(0);
-    }
-
-    @Test
-    public void testZeroEntities() {
-        // No listeners should be invoked with zero entities
-        TestEntitySetListener testListener = new TestEntitySetListener();
-        EntitySet entitySet = new EntitySet();
-        entitySet.addListener(testListener);
-        entitySet.addEntities(new Mask());
-        entitySet.removeEntities(new Mask());
-        testListener.verifyInserted(new int[0]);
-        testListener.verifyRemoved(new int[0]);
-        testListener.verifyInsertedInvocations(0);
-        testListener.verifyRemovedInvocations(0);
-    }
-
-    @Test
     public void testEquals() {
         EntitySet a = new EntitySet();
         EntitySet b = new EntitySet();
         assertEquals(a, b);
-        a.addListener(new TestEntitySetListener());
         assertEquals(a, b);
-        a.addEntity(0);
+        a.edit().addEntity(0);
         assertNotEquals(a, b);
-        a.addEntity(54);
+        a.edit().addEntity(54);
         assertNotEquals(a, b);
         assertNotEquals(a, new Object());
     }
@@ -181,11 +88,10 @@ public class EntitySetTest {
         EntitySet a = new EntitySet();
         EntitySet b = new EntitySet();
         assertEquals(a.hashCode(), b.hashCode());
-        a.addListener(new TestEntitySetListener());
         assertEquals(a.hashCode(), b.hashCode());
-        a.addEntity(0);
+        a.edit().addEntity(0);
         assertNotEquals(a.hashCode(), b.hashCode());
-        a.addEntity(54);
+        a.edit().addEntity(54);
         assertNotEquals(a.hashCode(), b.hashCode());
     }
 
@@ -193,36 +99,16 @@ public class EntitySetTest {
     public void testToString() {
         EntitySet set = new EntitySet();
         assertEquals("[]", set.toString());
-        set.addEntity(0);
+        set.edit().addEntity(0);
         assertEquals("[0]", set.toString());
-        set.addEntity(4);
+        set.edit().addEntity(4);
         assertEquals("[0, 4]", set.toString());
-        set.addEntity(2);
+        set.edit().addEntity(2);
         assertEquals("[0, 2, 4]", set.toString());
     }
 
     @Test(expected = RetinazerException.class)
-    public void testUnmodifiable0() {
-        new EntitySet().unmodifiable().addEntity(0);
-    }
-
-    @Test(expected = RetinazerException.class)
-    public void testUnmodifiable1() {
-        new EntitySet().unmodifiable().addEntities(new Mask());
-    }
-
-    @Test(expected = RetinazerException.class)
-    public void testUnmodifiable2() {
-        new EntitySet().unmodifiable().removeEntity(0);
-    }
-
-    @Test(expected = RetinazerException.class)
-    public void testUnmodifiable3() {
-        new EntitySet().unmodifiable().removeEntities(new Mask());
-    }
-
-    @Test(expected = RetinazerException.class)
-    public void testUnmodifiable4() {
-        new EntitySet().unmodifiable().clear();
+    public void testUnmodifiable() {
+        new EntitySet().view().edit().addEntity(0);
     }
 }