+++ CHANGELOG.md
@@ -20,21 +20,37 @@
 - `EntitySystem.setEngine(Engine)` is now `EntitySystem.setup()`
 - `Engine.getEntitiesFor(FamilyConfig)` is now `Engine.getFamily(FamilyConfig).getEntities()`
 - Reset method for `Engine` (`Engine.reset()`)
+- Enhanced `EntitySetListener` API
+  - Renamed to `EntityListener`
+  - `inserted(IntArray)` -> `inserted(EntitySet)`
+  - `removed(IntArray)` -> `removed(EntitySet)`
+- `Engine.addEntityListener(EntityListener)`
+- `Engine.removeEntityListener(EntityListener)`
+- `Family.addListener(EntityListener)`
+- `Family.removeListener(EntityListener)`
 - Enhanced `Mask` API
   - `Mask.getWord(int)`
   - `Mask.getWordCount()`
   - `Mask.set(long[])`
   - `Mask.setWord(int, long)`
   - `Mask.getWords()`
+  - `Mask.isEmpty()`
 - Revised `EntitySet` API
-  - `EntitySet.addEntities(IntArray)` removed
-  - `EntitySet.removeEntities(IntArray)` removed
-  - `EntitySet.addEntities(Mask)` checks existing entities
-  - `EntitySet.removeEntities(Mask)` checks existing entities
+  - Modification delegated to `EntitySetEdit`
+    - `EntitySet.addEntities(IntArray)` removed
+    - `EntitySet.removeEntities(IntArray)` removed
+    - `EntitySet.addEntities(Mask)` checks existing entities
+    - `EntitySet.removeEntities(Mask)` checks existing entities
+  - `EntitySet.edit()` returns `EntitySetEdit` or throws an exception if unmodifiable
+  - `EntitySet.unmodifiable()` renamed to `EntitySet.view()`
+  - `EntitySet.size()`
+  - `EntitySet.isEmpty()`
+  - Listeners are no longer supported, `EntityListener` is used for family events.
 - Support library for prototyping/jamming ("retinazer-beam")
   - `DependencyConfig` and `DependencyResolver` moved to retinazer-beam
 - `toString()` implementation for `EntitySet`
 - `RetinazerException` as a general-purpose runtime exception
+- Fix: EntitySet indices were invalidated even when there was no change
 - Fix: `IndexOutOfBoundsException` when negative index is used for `Bag.get`
 - Fix: No more global state, different engines running in different threads should work.